// src/contract/model/apiResponse.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ApiResponse {
  isSuccess?: boolean;
  data?: any | null;
  message?: string | null;
}