// src/contract/model/apiResponseOfAuthDto.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AuthDto } from './authDto';

export interface ApiResponseOfAuthDto {
  isSuccess?: boolean;
  data?: AuthDto | null;
  message?: string | null;
}